requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.0
httpx==0.27.0
SQLAlchemy==2.0.32
alembic==1.13.2
python-dotenv==1.0.1
apscheduler==3.10.4
pydantic==2.8.2